sct_extract_metric: Created API: extract_metric
